Print proxy env vars when applying authz resources
